Move study session view to viewMap
