+++ studySession/actions/index.js
@@ -1,4 +1,3 @@
-export switchToStudySession from './switchToStudySession';
 export initOrShuffle from './initOrShuffle';
 export move from './move';
 export flipActive from './flipActive';
