+++ studySession/constants.js
@@ -1,6 +1,4 @@
-export const VIEW='studySessionView';
 export const INIT_OR_SHUFFLE_DECKS='STUDY_SESSION_INIT_OR_SHUFFLE_DECKS';
-export const SWITCH_TO_STUDY_SESSION='STUDY_SESSION_SWITCH_TO_STUDY_SESSION';
 export const FLIP_ACTIVE='STUDY_SESSION_FLIP_ACTIVE';
 export const MOVE='STUDY_SESSION_MOVE';
 export const RED = 0;
