+++ viewMap/actions/index.js
@@ -1,1 +1,2 @@
 export switchToCreate from './switchToCreate';
+export switchToStudySession from './switchToStudySession';
