+++ viewMap/actions/switchToStudySession.js
