+++ viewMap/constants.js
@@ -1,2 +1,4 @@
 export const CREATE_VIEW = 'createCards';
 export const SWITCH_TO_CREATE = 'SWITCH_TO_CREATE';
+export const STUDY_SESSION_VIEW='studySessionView';
+export const SWITCH_TO_STUDY_SESSION='STUDY_SESSION_SWITCH_TO_STUDY_SESSION';
